+++ CTFd/api/v1/flags.py
@@ -7,11 +7,13 @@ from CTFd.api.v1.helpers.request import validate_args
 from CTFd.api.v1.helpers.schemas import sqlalchemy_to_pydantic
 from CTFd.api.v1.schemas import APIDetailedSuccessResponse, APIListSuccessResponse
 from CTFd.constants import RawEnum
-from CTFd.models import Flags, db
+from CTFd.models import Challenges, Flags, db
 from CTFd.plugins.flags import FLAG_CLASSES, get_flag_class
 from CTFd.schemas.flags import FlagSchema
-from CTFd.utils.decorators import admins_only,contributors_plus_admins_only
+from CTFd.utils.decorators import admins_only,contributors_contributors_plus_admins_only
 from CTFd.utils.helpers.models import build_model_filters
+from CTFd.utils.user import is_admin, is_contributor, is_contributor_plus
+from flask import session
 
 flags_namespace = Namespace("flags", description="Endpoint to retrieve Flags")
 
@@ -37,7 +39,7 @@ flags_namespace.schema_model(
 
 @flags_namespace.route("")
 class FlagList(Resource):
-    @contributors_plus_admins_only
+    @contributors_contributors_plus_admins_only
     @flags_namespace.doc(
         description="Endpoint to list Flag objects in bulk",
         responses={
@@ -68,7 +70,6 @@ class FlagList(Resource):
         q = query_args.pop("q", None)
         field = str(query_args.pop("field", None))
         filters = build_model_filters(model=Flags, query=q, field=field)
-
         flags = Flags.query.filter_by(**query_args).filter(*filters).all()
         schema = FlagSchema(many=True)
         response = schema.dump(flags)
@@ -77,7 +78,7 @@ class FlagList(Resource):
 
         return {"success": True, "data": response.data}
 
-    @contributors_plus_admins_only
+    @contributors_contributors_plus_admins_only
     @flags_namespace.doc(
         description="Endpoint to create a Flag object",
         responses={
@@ -97,18 +98,22 @@ class FlagList(Resource):
             return {"success": False, "errors": response.errors}, 400
 
         db.session.add(response.data)
-        db.session.commit()
+        
+        if is_admin() or is_contributor_plus() or (is_contributor() and response.data.challenge.author_id==session["id"]):
+            db.session.commit()
+            
 
-        response = schema.dump(response.data)
-        db.session.close()
+            response = schema.dump(response.data)
+            db.session.close()
 
-        return {"success": True, "data": response.data}
+            return {"success": True, "data": response.data}
+        return {"success": False}
 
 
 @flags_namespace.route("/types", defaults={"type_name": None})
 @flags_namespace.route("/types/<type_name>")
 class FlagTypes(Resource):
-    @contributors_plus_admins_only
+    @contributors_contributors_plus_admins_only
     def get(self, type_name):
         if type_name:
             flag_class = get_flag_class(type_name)
@@ -127,7 +132,7 @@ class FlagTypes(Resource):
 
 @flags_namespace.route("/<flag_id>")
 class Flag(Resource):
-    @contributors_plus_admins_only
+    @contributors_contributors_plus_admins_only
     @flags_namespace.doc(
         description="Endpoint to get a specific Flag object",
         responses={
@@ -150,21 +155,23 @@ class Flag(Resource):
 
         return {"success": True, "data": response.data}
 
-    @contributors_plus_admins_only
+    @contributors_contributors_plus_admins_only
     @flags_namespace.doc(
         description="Endpoint to delete a specific Flag object",
         responses={200: ("Success", "APISimpleSuccessResponse")},
     )
     def delete(self, flag_id):
         flag = Flags.query.filter_by(id=flag_id).first_or_404()
+        challenge = Challenges.query.filter_by(id=flag.challenge_id).first_or_404()
+        if is_admin() or is_contributor_plus() or (is_contributor() and challenge.author_id==session["id"]):
+            db.session.delete(flag)
+            db.session.commit()
+            db.session.close()
 
-        db.session.delete(flag)
-        db.session.commit()
-        db.session.close()
+            return {"success": True}
+        return{"sucess":False}
 
-        return {"success": True}
-
-    @contributors_plus_admins_only
+    @contributors_contributors_plus_admins_only
     @flags_namespace.doc(
         description="Endpoint to edit a specific Flag object",
         responses={
@@ -177,17 +184,20 @@ class Flag(Resource):
     )
     def patch(self, flag_id):
         flag = Flags.query.filter_by(id=flag_id).first_or_404()
-        schema = FlagSchema()
-        req = request.get_json()
+        challenge = Challenges.query.filter_by(id=flag.challenge_id).first_or_404()
+        if is_admin() or is_contributor_plus() or (is_contributor() and challenge.author_id==session["id"]):
+            schema = FlagSchema()
+            req = request.get_json()
 
-        response = schema.load(req, session=db.session, instance=flag, partial=True)
+            response = schema.load(req, session=db.session, instance=flag, partial=True)
 
-        if response.errors:
-            return {"success": False, "errors": response.errors}, 400
+            if response.errors:
+                return {"success": False, "errors": response.errors}, 400
 
-        db.session.commit()
+            db.session.commit()
 
-        response = schema.dump(response.data)
-        db.session.close()
+            response = schema.dump(response.data)
+            db.session.close()
 
-        return {"success": True, "data": response.data}
+            return {"success": True, "data": response.data}
+        return {"success": False}
